chore: update ui

+++ helpers/hexToString.js
@@ -2,7 +2,7 @@ export const stringToHex = (hex) => {
   let str = "";
   for (let i = 0; i < hex.length; i += 2) {
     const hexValue = hex.substr(i, 2);
-    const decimalValue = parseInt(hexValue, 16);
+    const decimalValue = parseFloat(hexValue, 16);
     str += String.fromCharCode(decimalValue);
   }
   return str;
